feat(adp): reorganize project settings
