+++ app/adp/settings/production.py
@@ -1,4 +1,4 @@
-from .settings import *
+from .base import *
 
 
 MIDDLEWARE = (
