+++ app/adp/settings/test.py
@@ -1,4 +1,4 @@
-from .settings import *
+from .base import *
 
 DATABASES = { 
     "default": {
